refactor(next-drupal): rearrange files into multiple entry points
Fixes #442
