+++ packages/next-drupal/src/navigation.ts
@@ -1,0 +1,1 @@
+export { useMenu } from "./navigation/use-menu"
