+++ packages/next-drupal/src/translation.ts
@@ -1,0 +1,4 @@
+export {
+  translatePath,
+  translatePathFromContext,
+} from "./translation/translate-path"
