+++ packages/next-drupal/src/utils/build-url.ts
@@ -1,0 +1,19 @@
+import { stringify } from "qs"
+
+export function buildUrl(
+  path: string,
+  params?: string | Record<string, string> | URLSearchParams
+): URL {
+  const url = new URL(
+    path.charAt(0) === "/"
+      ? `${process.env.NEXT_PUBLIC_DRUPAL_BASE_URL}${path}`
+      : path
+  )
+
+  if (params) {
+    // Use instead URLSearchParams for nested params.
+    url.search = stringify(params)
+  }
+
+  return url
+}
