+++ packages/next-drupal/src/utils/deserialize.ts
@@ -1,0 +1,9 @@
+import Jsona from "jsona"
+
+const dataFormatter = new Jsona()
+
+export function deserialize(body, options?) {
+  if (!body) return null
+
+  return dataFormatter.deserialize(body, options)
+}
