+++ packages/next-drupal/src/utils/sync-drupal-preview-routes.ts
@@ -1,0 +1,8 @@
+export function syncDrupalPreviewRoutes(path) {
+  if (window && window.top !== window.self) {
+    window.parent.postMessage(
+      { type: "NEXT_DRUPAL_ROUTE_SYNC", path },
+      process.env.NEXT_PUBLIC_DRUPAL_BASE_URL
+    )
+  }
+}
